requirements.txt
spacy==2.3.2
tqdm==4.51.0
pymongo==3.5.1
apache-airflow[celery]
apache-airflow[cryptography]
scikit-learn==0.23.2